支付宝URLScheme、注册页面字眼修改

+++ ZhiMaDi/UserCenter/Order/OrderCommentViewController.swift
@@ -233,8 +233,8 @@ class OrderCommentViewController: UIViewController,UITableViewDelegate,UITableVi
     //MARK: - MWPhotoBrowserDelegate
     func numberOfPhotosInPhotoBrowser(photoBrowser: MWPhotoBrowser!) -> UInt {
         let photos = self.photos[self.photoIndex]
-        return photos.count
-//        return UInt(photos.count)
+//        return photos.count
+        return UInt(photos.count)
     }
     func photoBrowser(photoBrowser: MWPhotoBrowser!, photoAtIndex index: UInt) -> MWPhotoProtocol! {
         let photos = self.photos[self.photoIndex]
